Fixed some missing info in AudioDB API

+++ MediaPortal/Source/Extensions/MetadataExtractors/OnlineLibraries/Libraries/AudioDbV1/Data/AudioDbArtist.cs
@@ -76,7 +76,7 @@ namespace MediaPortal.Extensions.OnlineLibraries.Libraries.AudioDbV1.Data
   public class AudioDbArtist
   {
     [DataMember(Name = "idArtist")]
-    public int ArtistId { get; set; }
+    public long ArtistId { get; set; }
 
     [DataMember(Name = "strArtist")]
     public string Artist { get; set; }
